moves pinia api plugin.
